separate termination criterion
